Moved the *Chain to *Manager
