+++ Resources/config/services.yml
@@ -1,9 +1,9 @@
 services:
-    dizda.cloudbackup.chain.database:
-        class: Dizda\CloudBackupBundle\Database\DatabaseChain
+    dizda.cloudbackup.manager.database:
+        class: Dizda\CloudBackupBundle\Manager\DatabaseManager
 
-    dizda.cloudbackup.chain.client:
-        class: Dizda\CloudBackupBundle\Client\ClientChain
+    dizda.cloudbackup.manager.client:
+        class: Dizda\CloudBackupBundle\Manager\ClientManager
 
     dizda.cloudbackup.client.cloudapp:
         class: Dizda\CloudBackupBundle\Client\CloudAppClient
@@ -84,8 +84,8 @@ services:
         class: Dizda\CloudBackupBundle\Manager\BackupManager
         arguments:
             - @logger
-            - @dizda.cloudbackup.chain.database
-            - @dizda.cloudbackup.chain.client
+            - @dizda.cloudbackup.manager.database
+            - @dizda.cloudbackup.manager.client
             - @dizda.cloudbackup.manager.processor
 
     dizda.cloudbackup.manager.processor:
